+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### [Unreleased]
 
-## [0.16.25] - 2021-06-08
+## [0.16.26] - 2021-06-08
 - optimize scheduling
+
+## [0.16.25] - 2021-05-30
+- implement average scoring
+- Fix TextEntry widget and scoring
 
 ## [0.16.24] - 2021-05-22
 - optimize scheduling